pyproject.toml
[project]
name = "database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "app",
    "fastparquet>=2024.11.0",
    "google-api-python-client>=2.185.0",
    "google-auth>=2.41.1",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-bigquery>=3.38.0",
    "gspread>=6.2.1",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pandas-gbq>=0.29.2",
    "pathlib>=1.0.1",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.2.11",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
]

[tool.uv.sources]
app = { workspace = true, editable = true }
